+++ cli/cli/commands/service/logs/logs.go
@@ -7,18 +7,15 @@ package logs
 
 import (
 	"context"
-	"github.com/docker/docker/client"
 	"github.com/docker/docker/pkg/stdcopy"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
-	docker_manager_types "github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager/types"
+	"github.com/kurtosis-tech/container-engine-lib/lib"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
-	labels_helper "github.com/kurtosis-tech/kurtosis-cli/cli/helpers/service_container_labels_by_enclave_id"
 	"github.com/kurtosis-tech/kurtosis-cli/commons/positional_arg_parser"
-	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
-	"io"
 	"strings"
 )
 
@@ -63,11 +60,13 @@ func run(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred parsing the positional args")
 	}
-	enclaveId := parsedPositionalArgs[enclaveIdArg]
-	guid := parsedPositionalArgs[guidArg]
+	enclaveIdStr := parsedPositionalArgs[enclaveIdArg]
+	enclaveId := enclave.EnclaveID(enclaveIdStr)
+	guidStr := parsedPositionalArgs[guidArg]
+	guid := service.ServiceGUID(guidStr)
 
 	// TODO Once KurtosisBackend can print logs
-	dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
+	/*dockerClient, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating the Docker client")
 	}
@@ -142,6 +141,48 @@ func run(cmd *cobra.Command, args []string) error {
 		}
 	}
 	// TODO ^^^^^^^^^^^^ Abstract everything below this point into KurtosisBackend when it's ready!!!! ^^^^^^^^^^^^^^^^
+	 */
+
+	kurtosisBackend, err := lib.GetLocalDockerKurtosisBackend()
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred getting local Docker Kurtosis backend")
+	}
+
+	userServiceFilters := &service.ServiceFilters{
+		GUIDs: map[service.ServiceGUID]bool{
+			guid: true,
+		},
+		EnclaveIDs: map[enclave.EnclaveID]bool{
+			enclaveId: true,
+		},
+	}
+
+	successfulUserServiceLogs, erroredUserServiceGuids, err := kurtosisBackend.GetUserServiceLogs(ctx, userServiceFilters, shouldFollowLogs)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred getting user service logs using filters '%+v'", userServiceFilters)
+	}
+
+	if len(erroredUserServiceGuids) > 0 {
+		err, found := erroredUserServiceGuids[guid]
+		if !found {
+			return stacktrace.NewError("Expected to find an error for user service with GUID '%v' on user service error map '%+v' but was not found; it should never happens, it is a bug in Kurtosis", guid, erroredUserServiceGuids)
+		}
+		return stacktrace.Propagate(err, "An error occurred getting user service logs for user service with GUID '%v'", guid)
+	}
+
+	userServiceReadCloserLog, found := successfulUserServiceLogs[guid]
+	if !found {
+		return stacktrace.NewError("Expected to find logs for user service with GUID '%v' on user service logs map '%+v' but was not found; it should never happens, it is a bug in Kurtosis", guid, userServiceReadCloserLog)
+	}
+
+	stdout := logrus.StandardLogger().Out
+	if _, err := stdcopy.StdCopy(stdout, stdout, userServiceReadCloserLog); err != nil {
+		return stacktrace.Propagate(
+			err,
+			"An error occurred copying the user service logs stream to STDOUT for user service with GUID '%v'",
+			guid,
+		)
+	}
 
 	return nil
 }