+++ cli/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Changes
+* Upgraded the container engine lib to x.x.x
 
 # 0.12.0
 ### Breaking Changes